# Optimize `Range#bsearch` by reducing `Integer#+` calls