--- a/benchmark/range_bsearch_bignum.yml
+++ b/benchmark/range_bsearch_bignum.yml
@@ -1,0 +1,10 @@
+prelude: |
+  first = 2**100
+  last = 2**1000
+  mid = (first + last) / 2
+  r = first..last
+
+benchmark:
+  first: r.bsearch { |x| x >= first }
+  mid: r.bsearch { |x| x >= mid }
+  last: r.bsearch { |x| x >= last }