This PR reduces `Integer#+` calls by using the implementation that requires fewer `+ 1` calculations.
This is particularly effective for Bignum ranges.

## benchmark

```yaml
prelude: |
  first = 2**100
  last = 2**1000
  mid = (first + last) / 2
  r = first..last

benchmark:
  first: r.bsearch { |x| x >= first }
  mid: r.bsearch { |x| x >= mid }
  last: r.bsearch { |x| x >= last }
```

|       |master|    PR|ratio|
|:------|-----:|-----:|----:|
|first  |3.089k|3.124k|1.011|
|mid    |2.066k|2.695k|1.304|
|last   |2.068k|2.714k|1.312|

In the case of Fixnum ranges, it seems to depend on the position of the target.
The same algorithm as in the Bignum case is adopted for clarity.

```yaml
prelude: |
  first = 1
  last = 10000
  mid = (first + last) / 2
  r = first..last

benchmark:
  first: r.bsearch { |x| x >= first }
  mid: r.bsearch { |x| x >= mid }
  last: r.bsearch { |x| x >= last }
```

|       |master|    PR|ratio|
|:------|-----:|-----:|----:|
|first  |2.437M|2.671M|1.096|
|mid    |2.618M|2.490M|0.951|
|last   |2.448M|2.476M|1.011|